Refactor project references and organize classes according to Clean Architecture
- project references fixed
- Middlewares moved into API project
- InitiateDb class moved to next to the ApplicationDbContext
- Orphan nuget in Infra project removed
- Project Services and Middleware registered
- Application Services are added to it own ServiceCollectionExtension class
- CustomSwaggerGen moved to extension folder in Api project
