+++ src/NadinSoft.CRUD.API/Extensions/SwaggerExtensions.cs
@@ -1,0 +1,37 @@
+namespace NadinSoft.CRUD.API.Extensions;
+
+public static class SwaggerExtensions
+{
+    public static void AddCustomSwaggerGen(this IServiceCollection service)
+    {
+        service.AddSwaggerGen(options =>
+        {
+            options.SwaggerDoc("v1", new() { Title = "ProductApp API", Version = "v1" });
+
+            options.AddSecurityDefinition("Bearer", new Microsoft.OpenApi.Models.OpenApiSecurityScheme
+            {
+                Name = "Authorization",
+                Type = Microsoft.OpenApi.Models.SecuritySchemeType.Http,
+                Scheme = "Bearer",
+                BearerFormat = "JWT",
+                In = Microsoft.OpenApi.Models.ParameterLocation.Header,
+                Description = "Enter your JWT token like: **Bearer YOUR_TOKEN_HERE**"
+            });
+
+            options.AddSecurityRequirement(new Microsoft.OpenApi.Models.OpenApiSecurityRequirement
+            {
+                {
+                    new Microsoft.OpenApi.Models.OpenApiSecurityScheme
+                    {
+                        Reference = new Microsoft.OpenApi.Models.OpenApiReference
+                        {
+                            Type = Microsoft.OpenApi.Models.ReferenceType.SecurityScheme,
+                            Id = "Bearer"
+                        }
+                    },
+                    []
+                }
+            });
+        });
+    }
+}
