+++ src/NadinSoft.CRUD.API/NadinSoft.CRUD.API.csproj
@@ -16,6 +16,7 @@
     </ItemGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\NadinSoft.CRUD.Application\NadinSoft.CRUD.Application.csproj" />
       <ProjectReference Include="..\NadinSoft.CRUD.Infrastructure\NadinSoft.CRUD.Infrastructure.csproj" />
     </ItemGroup>
