+++ src/NadinSoft.CRUD.Infrastructure/NadinSoft.CRUD.Infrastructure.csproj
@@ -24,7 +24,6 @@
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
-        <PackageReference Include="Swashbuckle.AspNetCore" Version="6.6.2"/>
     </ItemGroup>
 
 </Project>
